+++ pkg/services/secrets/database/database.go
@@ -126,7 +126,9 @@ func (ss *SecretsStoreImpl) ReEncryptDataKeys(
 	currProvider secrets.ProviderID,
 ) error {
 	keys := make([]*secrets.DataKey, 0)
-	if err := ss.sqlStore.NewSession(ctx).Table(dataKeysTable).Find(&keys); err != nil {
+	session := ss.sqlStore.NewSession(ctx)
+	defer session.Close()
+	if err := session.Table(dataKeysTable).Find(&keys); err != nil {
 		return err
 	}
