+++ pkg/services/secrets/migrator/reencrypt.go
@@ -17,8 +17,10 @@ func (s simpleSecret) reencrypt(ctx context.Context, secretsSrv *manager.Secrets
 		Id     int
 		Secret []byte
 	}
+	session := sqlStore.NewSession(ctx)
+	defer session.Close()
 
-	if err := sqlStore.NewSession(ctx).Table(s.tableName).Select(fmt.Sprintf("id, %s as secret", s.columnName)).Find(&rows); err != nil {
+	if err := session.Table(s.tableName).Select(fmt.Sprintf("id, %s as secret", s.columnName)).Find(&rows); err != nil {
 		logger.Warn("Could not find any secret to re-encrypt", "table", s.tableName)
 		return false
 	}
@@ -71,8 +73,10 @@ func (s b64Secret) reencrypt(ctx context.Context, secretsSrv *manager.SecretsSer
 		Id     int
 		Secret string
 	}
+	session := sqlStore.NewSession(ctx)
+	defer session.Close()
 
-	if err := sqlStore.NewSession(ctx).Table(s.tableName).Select(fmt.Sprintf("id, %s as secret", s.columnName)).Find(&rows); err != nil {
+	if err := session.Table(s.tableName).Select(fmt.Sprintf("id, %s as secret", s.columnName)).Find(&rows); err != nil {
 		logger.Warn("Could not find any secret to re-encrypt", "table", s.tableName)
 		return false
 	}
@@ -140,7 +144,10 @@ func (s jsonSecret) reencrypt(ctx context.Context, secretsSrv *manager.SecretsSe
 		SecureJsonData map[string][]byte
 	}
 
-	if err := sqlStore.NewSession(ctx).Table(s.tableName).Cols("id", "secure_json_data").Find(&rows); err != nil {
+	session := sqlStore.NewSession(ctx)
+	defer session.Close()
+
+	if err := session.Table(s.tableName).Cols("id", "secure_json_data").Find(&rows); err != nil {
 		logger.Warn("Could not find any secret to re-encrypt", "table", s.tableName)
 		return false
 	}
@@ -199,7 +206,11 @@ func (s alertingSecret) reencrypt(ctx context.Context, secretsSrv *manager.Secre
 	}
 
 	selectSQL := "SELECT id, alertmanager_configuration FROM alert_configuration"
-	if err := sqlStore.NewSession(ctx).SQL(selectSQL).Find(&results); err != nil {
+
+	session := sqlStore.NewSession(ctx)
+	defer session.Close()
+
+	if err := session.SQL(selectSQL).Find(&results); err != nil {
 		logger.Warn("Could not find any alert_configuration secret to re-encrypt")
 		return false
 	}
